requirements.txt
fastapi == 0.115.8
pydantic == 2.10.6
mysql-connector-python == 9.2.0
streamlit == 1.43.1
requests == 2.32.3
pandas == 2.2.3
pytest == 8.3.4
uvicorn == 0.34.0

